Import latest TF features and avoid tuple decomposition calls/RPCs. NOTE: The new TF bits triggers an ICE (in the Eigen library) on GCC 7.x and GCC 8.x! We are running out of compilers, as GCC 6.x was triggering another ICE before. Good news is that CLANG came to rescue. Using CLANG 7.x we are able to build:
sudo apt-get install clang-7 clang++-7
export CC=clang-7 CXX=clang++-7

We also made the TF framework build static, so we do not need to carry and extra .so.
This was necessary since the new TF bits were triggering a runtime error when the protobuf library was initializing, trying to define the same proto DB from both the XLA computation client, and the TF framework .so libraries.

+++ build_torch_xla_libs.sh
@@ -2,6 +2,8 @@
 
 set -ex
 
+CMD="${1:-install}"
+
 cd "$(dirname "$0")"
 PWD=`printf "%q\n" "$(pwd)"`
 BASE_DIR="$PWD"
@@ -11,13 +13,19 @@ THIRD_PARTY_DIR="$BASE_DIR/third_party"
 cp -r -f $THIRD_PARTY_DIR/xla_client $THIRD_PARTY_DIR/tensorflow/tensorflow/compiler/xla/
 
 pushd $THIRD_PARTY_DIR/tensorflow
-git reset --hard
-git clean -f
-bazel build -c opt //tensorflow/compiler/xla/xla_client:libxla_computation_client.so
-popd
+if [ "$CMD" == "install" ]; then
+  git reset --hard
+  git clean -f
+  bazel build --define framework_shared_object=false -c opt --cxxopt="-Wno-c++11-narrowing" \
+    //tensorflow/compiler/xla/xla_client:libxla_computation_client.so
 
-mkdir -p torch_xla/lib
-chmod 0644 $THIRD_PARTY_DIR/tensorflow/bazel-bin/tensorflow/compiler/xla/xla_client/libxla_computation_client.so
-cp $THIRD_PARTY_DIR/tensorflow/bazel-bin/tensorflow/compiler/xla/xla_client/libxla_computation_client.so torch_xla/lib
-chmod 0644 $THIRD_PARTY_DIR/tensorflow/bazel-bin/tensorflow/libtensorflow_framework.so
-cp $THIRD_PARTY_DIR/tensorflow/bazel-bin/tensorflow/libtensorflow_framework.so torch_xla/lib
+  popd
+  mkdir -p torch_xla/lib
+  chmod 0644 $THIRD_PARTY_DIR/tensorflow/bazel-bin/tensorflow/compiler/xla/xla_client/libxla_computation_client.so
+  cp $THIRD_PARTY_DIR/tensorflow/bazel-bin/tensorflow/compiler/xla/xla_client/libxla_computation_client.so torch_xla/lib
+elif [ "$CMD" == "clean" ]; then
+  bazel clean
+  popd
+else
+  popd
+fi
